backend/requirements.txt
# Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
alembic==1.12.0

# Machine Learning
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1
python-dateutil==2.8.2

# Security
bcrypt==4.0.1
cryptography==41.0.4
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# Caching and Session Management
redis==4.6.0
Flask-Session==0.5.0

# API and Serialization
marshmallow==3.20.1
webargs==8.3.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.6.2

# Development
flake8==6.1.0
black==23.9.1
isort==5.12.0
pre-commit==3.4.0

# Monitoring and Logging
sentry-sdk==1.32.0
structlog==23.1.0

# Background Tasks (optional)
celery==5.3.1
redis==4.6.0

# Time and Date
pytz==2023.3
arrow==1.2.3

# File Processing
Pillow==10.0.1
python-magic==0.4.27

# Environment and Configuration
pydantic==2.4.2
pydantic-settings==2.0.3

# HTTP Client
httpx==0.25.0

# Validation
email-validator==2.0.0
phonenumbers==8.13.20

# Financial Data (for future bank integration)
plaid-python==9.1.0

# Data Visualization (backend support)
plotly==5.16.1

# Performance Monitoring
py-spy==0.3.14